COURSES-444
This fix creates directory environments for Puppet courses via the
classrom module.
A production environment is created and an environment for each student.
Each student git puppetcode environment will have a manifests directory
for the environment site.pp and an environment.conf file to prevent
caching.
The production environment is linked to the normal mondules and manifest
directories, plus an environments.conf is created to prevent caching