Add rudimentary tuning class
* It can drop in a cron job to flush jruby instances every 30 minutes,
  on the advices of @cprice404 in https://tickets.puppetlabs.com/browse/PE-9704
* It can manage several tuning profiles for different levels of
  aggressiveness. To start with, only the profile used for the LVM is
  included.

Neither are enabled by default and the class isn't even included unless
we direct testers to try it out. I'd appreciate some benchmarks and
field testing, and I'd like @joshsamuelson to put together good numbers
to use for other tuning profiles.

After we have a better understanding of what the side effects may be, we
might include it by default.