DON'T MERGE: Recover CWD after pre-commit hook
I'm not completely certain why the stash & pop sometimes loses the CWD.
It has something to do with the order that files are created and
checked, etc. In any case, this *should* recover it.

Preliminary testing shows this to be effective. I will update this PR if
the problem has been completely resolved by end of class.