Don't merge: First pass at Shallow Gravy support
This should now theoretically support both Shallow Gravy and 3.8.x.
There were two main avenues for porting to the new filesystem layout.
First I parameterized `$codedir` and `$confdir` where I reasonably
could, so that if there are changes in the future, it should be
relatively easy to port. Second, I wrote a `classroom::compatibility`
shim class that drops in some janky symlinks so that legacy PE
filesystem layouts will work unchanged with the new layout. This means
that when it comes time to drop 3.x support, we just drop the shim.

This isn't 100% parameterized, because some of the paths exist in static
files, but the paths that would need corrected only exist in three
files:

* `files/pre-commit`
* `files/process_reports.rb`
* `files/sudoers.classroom`

I ended up conflating a lot of overall cleanups and code refactoring
into this, simply because I was touching the same lines of code and I
didn't want to write and rewrite the code multiple times.