Make containers composable
Breaking into a defined type instead of iterating over a hash means that
we can expose this type outside of these specific purposes. For example,
now it's trivial to *add* managed containers without modifying the code
of the classroom module.

This will likely be a large part of the capstone lab.