Restore management of modulepath symlink
In a demonstration of the dangers of hidden coupling, the removal of the
teams functionality that nobody uses accidentally removed the management
of the modulepath symlink. That *had* been in the team management, so
that it could link to either the student's own workdir or the team
workdir, depending on progress. It is now directly in the `workdir` type.

TODO: there is no longer any reason for `workdir` to be a type. That was
to allow a workdir for the student and for the team.

TRAINTECH-335 #resolved #time 30m