This corrects the permission problem
This allows the instructor to upload training material as the training
user. This should be all that remains to get the showoff/wk/pdftk stack
on every master. This has been tested with
fundamentals/practitioner/architect online. After we get a new VM build
out, it should also be tested offline to ensure that we've got all the
necessary packages cached.

TRAINTECH-660 #resolved #time 3h