WIP: can reuse puppet master's ssl certificates
This allows the presentation to use the same certificates as the Puppet
Console does. This means that the student would only need to accept a
single certificate for the classroom.

Don't merge yet, because for follow mode, or other interactivity, the
audience has to accept the certificate **permanently** and I'm not sure
that's good UX. I'm still looking for a way to get around that :-/