Release v1.7.0 (#310)
* Gitea requires the golang package

* Increasing timeout

When executing the Config ADDS line, it translates the timeout of 0 as 0, and assumes 50ms timeout, which causes the command to fail.  Upping to 1800s allows it to finish successfully.

* Update adserver.pp

* Missing previous directory

By default, WindowsPowerShell does not exist for any user.  It's the preferred directory for custom powershell modules (because of signing modules), but has to be created first.

* Added missing hierarchy level to get node-specific settings working

COURSES-2739 #resolved #time 15m

* Add sleep before acceptance tests

* Initial puppet run before classification

* inifile gem is required by vagrant

* MVP for virtual practitioner

We should probably refactor the support classes a little bit instead of
cutty pastey. But this is good enough for today. We'll refactor the
module quite a bit as we clean out some of the old cruft anyway.

The `classroom-control-vp` control repo does not yet exist.

* Basic spec test for pracitioner virtual course

* Add fixture of stubbed out pe module for rspec test

* Enable classroom PDF generation with courseware manager

This depends on https://github.com/puppetlabs/puppet-courseware-manager/pull/5

TRAINTECH-855 #time 30m

* Check for existence, not just directory

This is a file when filesync is in use.

TRAINTECH-1062 #resolved #time 15m

* jeeze

Ok, this will get actual somewhat meaningful revisions

* cmake is now a dependency

because we pinned to a known good version, commonmarker is now a hard dependency of showoff. It requires cmake to compile the native libcmark code.

* Staging dir = codedir for hiera course

* Staging dir = codedir for hiera course

* Put reporting tools on VP nodes

TRAINTECH-1061 #resolved #time 15m

* on the master only

* Manage hiera & hiera_explain on all virtual nodes

This unifies how Hiera is managed for all classes, virtual and not. It
also puts the `hiera_explain.rb` script on all nodes for easier
debugging when running it locally. (currently only the case in
practitioner)

TRAINTECH-1087 #resolved #time 45m

* Add openssl to agent nodes

This creates a new grab-bag class for misc. packages to be installed on
the agent. The `openssl` package is already cached on the master, so
this installs just fine.

TRAINTECH-1068 #resolved #time 30m

* Added support for offline mode for virtual courses and configuration
of the gitea Git service on the master

* Fix powershell to find all Ethernet interfaces

* Kerberos fixes for augeas (#302)

* Kerberos fixes for augeas

* Manage the lens

* Update init.pp

* Update virtual.pp

* Update augeas.pp

* Create lens directories

* Include facts for augeas

* Pass the offline parameter to the classroom::virtual class so
we can test the value correctly

* Enabled the puppetfactory gitea plugin for courses that operate in
offline mode with the Gitea service enabled on the master

* Only enable the Gitea puppetfactory plugin when operating in offline
mode

* Added environment

COURSES-2630 #resolved #time 30m

* Use ipv4 for postfix

If this isn't set, postfix will default to "all" which tries to use ipv6 even if ipv6 isn't enabled on the system.

* Disable ipv6 on postfix for all agents

* Make control_owner required to remind instructor to clone the
control repository to their own namespace

* Version bump 1.7.0

* The control_owner parameter is now required to omit the test
that does not specify it.

* Remove the test that omits setting control_owner, because it is
required now

* Fix the spec test for VP (#308)

* Enabled and tested DSC resources (#301)

* Pin to a known working version of choco (#309)

TRAINTECH-1112 #resolved #time 30m

* Put peadmin on practitioner agents too (#288)

We'll want to test this before committing to it.