WIP: only set .ssh permissions on Linux
Ethan suggested not setting these permissions on Windows, thinking that it might be affecting the key permissions. I'm not convinced, because I think ssh will complain about permissions. Still, we should test it and find out.

If it does, then maybe we need to be using the acl module instead?