(TRAINTECH-1168): Fix adserver configuration
TRAINTECH-1168 #resolved #time 4d

These updates help us automatically set up a machine as an adserver for
the windows essentials course.

Changes that were made include making the reboot after setting up the AD
domain be an immediate reboot, as opposed to a reboot when pending (as
noted by DSC).

This change also installs a known-good version of powershell5 using
chocolatey instead of running a script to install WMF5.

You can change the version we pin. This is just the latest version that is known-good.