We don't need to disable timeout anymore
Both the deploy button and the post-update properly clear the
environment now. The only time this might have an affect is if
instructors edit code live. (which they should not be doing)

Besides, in my tests, the Pe_ini_setting flipping this back to
'unlimited' always seemed to win anyway.