Add a deployment canary
This simply attempts a deployment until it succeeds. Because the PE module isn't idempotent, and because we have to run Puppet to enable code manager in the first place, and because it depends on the Git server being up, this will probably fail a few times. So we ignore failures and just keep trying.