Removes the non-gitea options
We've consolidated to using Gitea across the board. This codifies that
and removes the number of variations we need to test & validate. It also
simplifies the experience for the instructor, since they're not
presented with unclear configuration options.

This also adds a spec for puppetize, since its configuration is slightly
different from the other virtuals.

TRAINTECH-1445 #resolved #time 2h