Windows spec tests for PI (incomplete)
This adds all the fixtures and framework for running spec tests on
windows to validate our classes. Unfortunately, the chocolateyfeature
type and the powershell exec provider don't test cleanly, so we can't
actually turn this on. I commented out those parts and used this spec to
validate everything else (and corrected a couple issues), but ultimately
had to comment the spec out until such a time as we can validate the
above mentioned type & provider.

Ethan is working on some recommendations, so we can correct it later,
but this PR is safe to merge as is, since the spec is commented out.
Notes about unclean tests are in the source.

TRAINTECH-1504 #time 2h