Switch to Virtual First
This switches to virtual first delivery for our materials. The existing
ILT courseware is moved to Legacy content. The Virtual courseware is
promoted to being *the way* that we deliver courses, with two variants.

The only difference between local/virtual deliveries right now is the
one slide about introducing the virtual platform. But maintaining the
two variants means that we can re-chunk as needed to meet the delivery
requirements.

This should be released on Wednesday or Thursday, so no trainings are
affected.

TRAINTECH-1563 #resolved #time 1h